requirements.txt
beautifulsoup4==4.12.3
EbookLib==0.18
elevenlabs==1.50.3
Flask==3.1.0
openai==1.59.3
pydub==0.25.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.2
Werkzeug==3.1.3
